# spec/openapi-edukate-backend.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:5810
security:
- cookieAuth: []
...
paths:
  /api/v1/submissions:
    post:
      tags:
      - Submissions
      summary: Upload a submission
      description: Creates a new submission for a problem with the provided files
      operationId: uploadSubmission
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionRequest"
        required: true
      responses:
        "200":
          description: Successfully created submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionDto"
        "403":
          description: Access denied - No submit permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionDto"
        "404":
          description: "User, problem, or files not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionDto"
      security:
      - cookieAuth: []
...
  /api/v1/submissions/{bookSlug}/{code}/{username}:
    get:
      tags:
      - Submissions
      summary: Get submissions by username and problem key
      description: Retrieves paginated submissions for a specific user and problem
      operationId: getSubmissionsByUsernameAndProblemKey
      parameters:
      - name: bookSlug
        in: path
        description: Book slug
        required: true
        schema:
          type: string
      - name: code
        in: path
        description: Problem code
        required: true
        schema:
          type: string
      - name: username
        in: path
        description: Username
        required: true
        schema:
          type: string
      - name: page
        in: query
        description: Page number (zero-based)
        required: false
        schema:
          type: string
          minimum: 0
      - name: size
        in: query
        description: Number of submissions per page
        required: false
        schema:
          type: string
          exclusiveMinimum: 0
          maximum: 100
          minimum: 1
      responses:
        "200":
          description: Successfully retrieved submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubmissionDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubmissionDto"
        "403":
          description: Access denied - Requires MODERATOR role
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubmissionDto"
  /api/v1/submissions/my:
    get:
      tags:
      - Submissions
      summary: Get my submissions
      description: "Requires an authenticated session via the gateway cookie. Returns\
        \ the authenticated user's submissions, optionally filtered by problemId.\
        \ Results are paginated (page, size) and sorted by createdAt in descending\
        \ order."
      operationId: getMySubmissions
      parameters:
      - name: problemKey
        in: query
        description: Optional problem key (bookSlug/code) to filter the user's submissions
        required: false
        schema:
          type: string
      - name: page
        in: query
        description: Page number (zero-based)
        required: false
        schema:
          type: string
          minimum: 0
      - name: size
        in: query
        description: Number of submissions per page (max 100)
        required: false
        schema:
          type: string
          exclusiveMinimum: 0
          maximum: 100
          minimum: 1
      responses:
        "200":
          description: Successfully retrieved submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubmissionDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubmissionDto"
      security:
      - cookieAuth: []
  /api/v1/submissions/by-id/{id}:
    get:
      tags:
      - Submissions
      summary: Get submission by ID
      description: Retrieves a specific submission by its ID for the authenticated
        user
      operationId: getSubmissionById
      parameters:
      - name: id
        in: path
        description: Submission ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successfully retrieved submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionDto"
        "403":
          description: Access denied - user id does not match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionDto"
        "404":
          description: Submission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionDto"
      security:
      - cookieAuth: []
...
components:
  schemas:
    CreateSubmissionRequest:
      type: object
      properties:
        problemKey:
          type: string
          minLength: 1
        fileNames:
          type: array
          items:
            type: string
          minItems: 1
      required:
      - fileNames
      - problemKey
    SubmissionDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        problemKey:
          type: string
        userName:
          type: string
        status:
          type: string
          enum:
          - PENDING
          - FAILED
          - SUCCESS
        createdAt:
          type: string
          format: date-time
        fileUrls:
          type: array
          items:
            type: string
      required:
      - createdAt
      - fileUrls
      - id
      - problemKey
      - status
      - userName
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: JWT authentication cookie
      name: X-Auth
      in: cookie